adjust repository structure
